Evaluation of pairwise cell relationships, conditional on a 3rd population.

Kontextual identifies the relationship between two cell types which are conditional on the spatial behaviour of a 3rd cell population, for a particular radius (r).

Usage

Kontextual(
  cells,
  parentDf = NULL,
  r,
  from = NULL,
  to = NULL,
  parent = NULL,
  inhom = TRUE,
  edgeCorrect = FALSE,
  window = "convex",
  window.length = NA,
  weightQuantile = 0.8,
  includeZeroCells = TRUE,
  includeOriginal = TRUE,
  spatialCoords = c("x", "y"),
  cellType = "cellType",
  imageID = "imageID",
  cores = 1
)

Arguments

cells

A SingleCellExperiment, SpatialExperiment or a list of data.frames containing columns specifying the imageID, cellType, and x and y spatial coordinates.

parentDf

A data frame from parentCombinations

r

Radius to evaluated pairwise relationships between from and to cells.

from

The first cell type to be evaluated in the pairwise relationship.

to

The second cell type to be evaluated in the pairwise relationship.

parent

The parent population of the from cell type (must include from cell type).

inhom

A logical value indicating whether to account for inhomogeneity.

edgeCorrect

A logical value indicating whether to perform edge correction.

window

Type of window for data, either `square`, `convex` or `concave`, passed into makeWindow

window.length

A tuning parameter for controlling the level of concavity when estimating concave windows. Passed into makeWindow

weightQuantile

A decimal value indicating what quantile of parent density used to weight the `from` cells.

includeZeroCells

A logical value indicating whether to include cells with zero counts in the pairwise association calculation.

includeOriginal

A logical value to return the original L function values along with the kontextual values.

spatialCoords

The columns which contain the x and y spatial coordinates.

cellType

The column which contains the cell types.

imageID

The column which contains image identifiers.

cores

Number of cores for parallel processing.

Value

A Koditional result object

Examples

# Load data
data("headSCE")

CD4_Kontextual <- Kontextual(
  cells = headSCE,
  r = 50,
  from = "TC_CD4",
  to = "SC5",
  parent = c("TC_CD4", "TC_CD8"),
  cores = 40
)

head(CD4_Kontextual)
#>   imageID        test  original kontextual  r weightQuantile inhom  edge
#> 1       1 TC_CD4__SC5 -11.50459    -1.94470 50            0.8  TRUE FALSE
#> 2       5 TC_CD4__SC5 -15.07237     1.37433 50            0.8  TRUE FALSE
#> 3       6 TC_CD4__SC5 -16.45579    12.28765 50            0.8  TRUE FALSE
#> 4       7 TC_CD4__SC5 -17.68033   -11.51087 50            0.8  TRUE FALSE
#> 5       8 TC_CD4__SC5 -15.78141    11.89574 50            0.8  TRUE FALSE
#>   includeZeroCells window window.length
#> 1             TRUE convex            NA
#> 2             TRUE convex            NA
#> 3             TRUE convex            NA
#> 4             TRUE convex            NA
#> 5             TRUE convex            NA